feat(git): use rebase strategy for `git pull`
Since git 2.27.0 you will get a nagging message complaining if not
strategy is configure of how to deal with diverging branches on pulls.
The options are to set `pull.rebase=false` (merge, default),
`pull.rebase=true` (rebase), or `pull.ff=only` (fast-forward only).

Chose `pull.rebase=true` despite this potentially being a dangerous
operation as it could rewrite history which is a bad idea if it is
already published.  However that is exceedingly unlikely for me.  Also
can override this on a per-repository setting if it ever becomes
relevant.