feat(alacritty): bring back scrollback; force TERM
It's helpful to have some scrollback by default instead of always
relying on running `tmux`.

Also it looks like that `alacritty` is not setting `TERM` to `alacritty`
as it claims in the documentation if the termcap info is not in a global
location.  That is likely reasonable as picking up the correct location
relies on the `TERMINFO` env variable which only gets set by the shell.

Using `TERM=alacritty` does seem to break `vim` (`neovim` seems fine),
needs some more investigation, but since I (currently) default to
`neovim` it's not a high priority to fix.