fix(tmux): improve background handling
Theoretically this is meant to not draw a background colour in tmux so
that it works with transparency in terminals that applied it only to the
default background colour.  However this doesn't seem to quite work
(nonetheless, be explicit to use the default colours).

Also enable true colour mode for alacritty, this fixes differing
background colours between just alacritty and tmux in alacritty (which
has a slightly redish tint otherwise).