feat(bash): filetype detection for bash configs
Vim has various flavours of syntax highlighting for shell files.
It will automatically use bash for files it knows are bash, which are
the main top-level config files (.bashrc and friends) and files with
ending `.bash`.

Previously tried to achieve the correct highlight of the bash
configuration files via a modeline, but that only set Bourne shell
(unless a global configuration in vim is set to assume bash by default).
Much simpler and more reliable to drop the modeline and let it
auto-detect the right setting by just using a `.bash` suffix for the
files.

For more information see `:help ft-bash-syntax`.