fix(starship,bash): use self-build starship
Starship removed the ability to use `` in the prompt to force an empty
output, I am using that to get a prompt on an empty new line.  The
locally built starship version removes the 'fix' that started to
escape `` so that it is still evaluated by bash in my version.