feat(bash): build rust programs for local machine
Set an environment variable so that compiling Rust code will optimise
the binary for the local machine.  Note: this is not suitable for
producing binaries for distribution, but I am not doing that but rather
build things locally that then may just as well be best-adapted to where
they are built.