feat(sway,i3status-rust): improve status bar in sway
Use `i3status-rust` to generate the status bar now, it is compatible
with `sway`.  Also don't use absolute paths in the sway config when
referring to binaries/config files.  It does not seem to understand
$XDG_CONFIG_HOME so currently still do that manually with $HOME/.config.

Switch to using IBM Plex Sans instead of -Mono for the status bar.  This
is more efficient with the space and it isn't sprawling too wide.