fix(ssh): provide a way to have local additions
I cannot use `UseKeychain yes` in a config file that is meant for Linux
(it is an invalid key on that type and thus it fails to load the config
file), but need to use it on Mac.  So instead `Include` a local config
that can be used to set these things while still reusing the common
bits.