feat(vim,conda): update LSP setup
Firstly, `vim-lsp` has been renamed to `vim-lspconfig`, so fix-up the
submodule name/path/et cetera.  Then add some more useful keybindings
for LSP things.

Finally, `python-language-server` is not maintained by Microsoft any
more and replaced with a proprietary solution only available in VSCode.
So switching to a fork of `python-language-server` that is still
maintained instead.  Also adding in a couple of neat plug-ins to make
it more useful (in particular the `mypy` integration).