feat(vim): add obsidian plugin
This allows to write notes more comfortably from neovim instead of in
Obsidian.  It also allows to follow links and at least works vaguely
okay of the daily notes as well.  Will see how I am getting on with
this, it does pull in a couple of meaty other plugins (and I haven't
even added the suggested one for completions yet).